Almost completed backend, only need some login features

+++ Backend/src/main/java/com/revature/controllers/UserController.java
@@ -2,6 +2,7 @@ package com.revature.controllers;
 
 import com.revature.models.User;
 import com.revature.services.UserService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -26,18 +27,44 @@ public class UserController {
 
 
     @GetMapping
-    public List<User> getAllUsers() {
-        return us.getAllUsers();
+    public ResponseEntity <Object> getAllUsers( HttpSession session) {
+        String role = (String) session.getAttribute("role");
+        if (role ==null || !role.equalsIgnoreCase("Manager")) {
+            return ResponseEntity.status(403).body("Unauthorized access: Only managers can view all users");
+        }
+
+        if (us.getAllUsers() == null) {
+            return ResponseEntity.status(400).body("No users found");
+        }
+
+        return ResponseEntity.ok(us.getAllUsers());
     }
 
     @DeleteMapping("/{userId}")
-    public ResponseEntity<Object> deleteUserById(@PathVariable int userId) {
+    public ResponseEntity<Object> deleteUserById(@PathVariable int userId, HttpSession session) {
+        String role = (String) session.getAttribute("role");
+        if (role ==null || !role.equalsIgnoreCase("Manager")) {
+            return ResponseEntity.status(403).body("Unauthorized access: Only managers can delete users");
+        }
+
+        if (us.getUserById(userId) == null) {
+            return ResponseEntity.status(400).body("User not found with ID: " + userId);
+        }
         this.us.deleteUserById(userId);
         return ResponseEntity.ok("User with ID: " + userId + " deleted successfully");
     }
 
     @PatchMapping("/{userId}")
-    public ResponseEntity<Object> updateUserRole(@PathVariable int userId, @RequestBody String role) {
+    public ResponseEntity<Object> updateUserRole(@PathVariable int userId, @RequestBody String role, HttpSession session) {
+        String role2 = (String) session.getAttribute("role");
+        if (role2 ==null || !role2.equalsIgnoreCase("Manager")) {
+            return ResponseEntity.status(403).body("Unauthorized access: Only managers can update user roles");
+        }
+
+        if (us.getUserById(userId) == null) {
+            return ResponseEntity.status(400).body("User not found with ID: " + userId);
+        }
+
         User updatedUser = this.us.updateUserRole(userId, role);
         return updatedUser == null ? ResponseEntity.status(400).body("User not found with ID: " + userId) : ResponseEntity.ok(updatedUser);
     }
